Move errors and exceptions to an exceptions module.
